libfinaliza s/ bonus

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: fsuomins <fsuomins@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 23:07:05 by fsuomins          #+#    #+#             */
-/*   Updated: 2022/09/14 22:11:27 by fsuomins         ###   ########.fr       */
+/*   Updated: 2022/09/22 22:44:50 by fsuomins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	c;
+	size_t	c;
 
 	c = 0;
 	while (str[c] != '\0')
